requirements.txt
dj-database-url==0.3.0
Django==1.8
django-cors-headers==1.1.0
django-filter==0.10.0
djangorestframework==3.1.1
newrelic==2.52.0.40
pep8==1.5.7
psycopg2==2.6
waitress==0.8.9
wheel==0.24.0
